local-ai-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
aiohttp==3.9.1
python-dotenv==1.0.0
GitPython==3.1.40
docker==7.0.0
pydantic==2.5.3
